+++ puma/tests/hlplots/test_n_track_origin.py
@@ -3,6 +3,9 @@
 from __future__ import annotations
 
 import os
+
+# Uncomment line below to update expected image
+# import shutil
 import tempfile
 import unittest
 from urllib.request import urlretrieve
@@ -132,7 +135,7 @@ class NTrackOriginTestCase(unittest.TestCase):
                 )
             )
 
-    def test_extra_kwargs(self):
+    def test_extra_kwargs_all_flavour(self):
         """Test extra kwargs."""
         n_tracks_per_origin(
             flavour_list=[
@@ -176,6 +179,65 @@ class NTrackOriginTestCase(unittest.TestCase):
             )
         )
 
+    def test_extra_kwargs_all_samples(self):
+        """Test extra kwargs."""
+        n_tracks_per_origin(
+            flavour_list=[
+                Flavours.bjets,
+                Flavours.cjets,
+                Flavours.ujets,
+            ],
+            files={
+                "ttbar_1": {
+                    "filepath": self.r22_test_file,
+                    "n_jets": 10000,
+                    "tracks_name": "tracks_loose",
+                    "pt_bins": np.linspace(20_000, 100_000, 20),
+                    "process_label": "$t\\bar{t}$ 1",
+                    "jet_pt_variable": "pt_btagJes",
+                    "track_truth_variable": "truthOriginLabel",
+                    "flavour_label_variable": "HadronConeExclTruthLabelID",
+                },
+                "ttbar_2": {
+                    "filepath": self.r22_test_file,
+                    "n_jets": 10000,
+                    "tracks_name": "tracks_loose",
+                    "pt_bins": np.linspace(100_000, 250_000, 20),
+                    "process_label": "$t\\bar{t}$ 2",
+                    "jet_pt_variable": "pt_btagJes",
+                    "track_truth_variable": "truthOriginLabel",
+                    "flavour_label_variable": "HadronConeExclTruthLabelID",
+                },
+            },
+            plot_path=self.actual_plots_dir,
+            plot_type="all_samples_one_flavour",
+            plot_name="extra_kwargs",
+            track_origin_dict=None,
+            plot_format="png",
+            atlas_second_tag="Test extra kwargs",
+        )
+
+        # Define a plot name
+        names = [
+            "extra_kwargs_bjets_all_samples.png",
+            "extra_kwargs_cjets_all_samples.png",
+            "extra_kwargs_ujets_all_samples.png",
+        ]
+
+        # Check all plots
+        for name in names:
+            # Uncomment line below to update expected image
+            # shutil.copy(f"{self.actual_plots_dir}/{name}", f"{self.expected_plots_dir}/{name}")
+
+            # Check
+            self.assertIsNone(
+                compare_images(
+                    f"{self.actual_plots_dir}/{name}",
+                    f"{self.expected_plots_dir}/{name}",
+                    tol=1,
+                )
+            )
+
     def test_custom_track_origins(self):
         """Test custom grouping of track origins."""
         n_tracks_per_origin(
@@ -212,9 +274,7 @@ class NTrackOriginTestCase(unittest.TestCase):
         name = "ttbar_test_custom_track_origins_all_flavour.png"
 
         # Uncomment line below to update expected image
-        # shutil.copy(
-        #     f"{self.actual_plots_dir}/{name}", f"{self.expected_plots_dir}/{name}"
-        # )
+        # shutil.copy(f"{self.actual_plots_dir}/{name}", f"{self.expected_plots_dir}/{name}")
 
         # Check
         self.assertIsNone(
@@ -224,3 +284,30 @@ class NTrackOriginTestCase(unittest.TestCase):
                 tol=1,
             )
         )
+
+    def test_wrong_plot_type(self):
+        """Test wrong plot type error raising."""
+        with self.assertRaises(ValueError):
+            n_tracks_per_origin(
+                flavour_list=[
+                    Flavours.bjets,
+                    Flavours.cjets,
+                    Flavours.ujets,
+                ],
+                files={
+                    "ttbar_test_custom_track_origins": {
+                        "filepath": self.r22_test_file,
+                        "n_jets": 10000,
+                        "tracks_name": "tracks_loose",
+                        "pt_bins": np.linspace(20_000, 250_000, 20),
+                        "process_label": "$t\\bar{t}$",
+                        "jet_pt_variable": "pt_btagJes",
+                        "track_truth_variable": "truthOriginLabel",
+                        "flavour_label_variable": "HadronConeExclTruthLabelID",
+                    },
+                },
+                plot_path=self.actual_plots_dir,
+                plot_type="crash",
+                track_origin_dict=None,
+                plot_format="png",
+            )